% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iggASVProps}
\alias{iggASVProps}
\title{IgG Proportions}
\format{
## `iggASVProps`
A data frame:
\describe{
  \item{Taxa}{The relevant ASV}
  \item{Populations}{Four population columns}
}
}
\usage{
iggASVProps
}
\description{
IgG Proportions
}
\keyword{datasets}
