% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iggFracs}
\alias{iggFracs}
\title{IgG Fractions}
\format{
## `iggFracs`
A vector containing fractions applied to each of the population columns in order
}
\usage{
iggFracs
}
\description{
IgG Fractions
}
\keyword{datasets}
