% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{igmASVProps}
\alias{igmASVProps}
\title{IgM Proportions}
\format{
## `igmASVProps`
A data frame:
\describe{
  \item{Taxa}{The relevant ASV}
  \item{Populations}{Four population columns}
}
}
\usage{
igmASVProps
}
\description{
IgM Proportions
}
\keyword{datasets}
