% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{igmFracs}
\alias{igmFracs}
\title{IgM Fractions}
\format{
## `igmFracs`
A vector containing fractions applied to each of the population columns in order
}
\usage{
igmFracs
}
\description{
IgM Fractions
}
\keyword{datasets}
