% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mflow.R
\name{mflow_apply}
\alias{mflow_apply}
\title{mFLOW-Seq probabilities for ASV data in matrix or data frame format}
\usage{
mflow_apply(asv_matrix, fraction_abundances)
}
\arguments{
\item{asv_matrix}{Proportion of ASVs in each sample. ASVs should be in rows,
the mFLOW-Seq channels should be in columns. The whole matrix should
correspond to one specimen that was sorted and sequenced.}

\item{fraction_abundances}{Total abundance for each mFLOW channel}
}
\value{
A matrix of probabilities that each ASV was tagged in each channel.
}
\description{
mFLOW-Seq probabilities for ASV data in matrix or data frame format
}
