% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mflow.R
\name{mflow_probability}
\alias{mflow_probability}
\title{Probability of an ASV appearing in each channel from mFLOW-Seq}
\usage{
mflow_probability(asv_proportions, fraction_abundances)
}
\arguments{
\item{asv_proportions}{Proportion of ASVs in sample}

\item{fraction_abundances}{Total abundance for each mFLOW channel}
}
\value{
A vector of probabilities that the ASV was tagged in each channel.
}
\description{
This is the function you would use for data in long format, operating on one
specimen and one ASV at a time
}
