% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mflow.R
\name{normalize_sum}
\alias{normalize_sum}
\title{Ensure that the sum of a vector equals a given total}
\usage{
normalize_sum(x, total = 1)
}
\arguments{
\item{x}{A numeric vector}

\item{total}{The desired sum}
}
\value{
The original vector scaled to sum to the given total
}
\description{
Ensure that the sum of a vector equals a given total
}
